Moved manifest clone to robottelo.utils.manifest
