+++ robottelo/host_helpers/cli_factory.py
@@ -29,7 +29,7 @@ from robottelo.cli.base import CLIReturnCodeError
 from robottelo.cli.proxy import CapsuleTunnelError
 from robottelo.config import settings
 from robottelo.host_helpers.repository_mixins import initiate_repo_helpers
-from robottelo.utils import clone
+from robottelo.utils.manifest import clone
 
 
 class CLIFactoryError(Exception):
